numeric extension move into core part 3
moved numeric/convolve.hpp into core
