+++ example/convolve2d.cpp
@@ -1,15 +1,16 @@
-// 
+//
 // Copyright 2019 Miral Shah <miralshah2211@gmail.com>
-// 
-// Distributed under the Boost Software License, Version 1.0 
-// See accompanying file LICENSE_1_0.txt or copy at 
-// http://www.boost.org/LICENSE_1_0.txt 
-// 
+// Copyright 2019 Mateusz Loskot <mateusz at loskot dot net>
+// Copyright 2021 Pranam Lashkari <plashkari628@gmail.com>
+//
+// Use, modification and distribution are subject to the Boost Software License,
+// Version 1.0. (See accompanying file LICENSE_1_0.txt or copy at
+// http://www.boost.org/LICENSE_1_0.txt)
 
 #include <vector>
 #include <iostream>
 #include <boost/gil/extension/numeric/kernel.hpp>
-#include <boost/gil/extension/numeric/convolve.hpp>
+#include <boost/gil/image_processing/convolve.hpp>
 #include <boost/gil/extension/io/png.hpp>
 
 #include <boost/gil/extension/io/jpeg.hpp>
