+++ example/harris.cpp
@@ -1,5 +1,6 @@
 //
 // Copyright 2019 Olzhas Zhumabek <anonymous.from.applecity@gmail.com>
+// Copyright 2021 Pranam Lashkari <plashkari628@gmail.com>
 //
 // Use, modification and distribution are subject to the Boost Software License,
 // Version 1.0. (See accompanying file LICENSE_1_0.txt or copy at
@@ -7,8 +8,8 @@
 //
 
 #include <boost/gil/extension/io/png.hpp>
-#include <boost/gil/extension/numeric/convolve.hpp>
 #include <boost/gil/image.hpp>
+#include <boost/gil/image_processing/convolve.hpp>
 #include <boost/gil/image_processing/harris.hpp>
 #include <boost/gil/image_processing/numeric.hpp>
 #include <boost/gil/image_view.hpp>
