+++ example/sobel_scharr.cpp
@@ -1,15 +1,15 @@
-// 
-// // Copyright 2019 Olzhas Zhumabek <anonymous.from.applecity@gmail.com>
-// 
-// Distributed under the Boost Software License, Version 1.0 
-// See accompanying file LICENSE_1_0.txt or copy at 
-// http://www.boost.org/LICENSE_1_0.txt 
 //
+// Copyright 2019 Olzhas Zhumabek <anonymous.from.applecity@gmail.com>
+// Copyright 2021 Pranam Lashkari <plashkari628@gmail.com>
+//
+// Distributed under the Boost Software License, Version 1.0
+// See accompanying file LICENSE_1_0.txt or copy at
+// http://www.boost.org/LICENSE_1_0.txt
 
 #include <boost/gil/typedefs.hpp>
 #include <boost/gil/image_processing/numeric.hpp>
 #include <boost/gil/extension/io/png.hpp>
-#include <boost/gil/extension/numeric/convolve.hpp>
+#include <boost/gil/image_processing/convolve.hpp>
 #include <string>
 #include <iostream>
