+++ include/boost/gil/image_processing/threshold.hpp
@@ -1,5 +1,6 @@
 //
 // Copyright 2019 Miral Shah <miralshah2211@gmail.com>
+// Copyright 2021 Pranam Lashkari <plashkari628@gmail.com>
 //
 // Use, modification and distribution are subject to the Boost Software License,
 // Version 1.0. (See accompanying file LICENSE_1_0.txt or copy at
@@ -20,7 +21,7 @@
 
 #include <boost/gil/image.hpp>
 #include <boost/gil/extension/numeric/kernel.hpp>
-#include <boost/gil/extension/numeric/convolve.hpp>
+#include <boost/gil/image_processing/convolve.hpp>
 #include <boost/gil/image_processing/numeric.hpp>
 
 namespace boost { namespace gil {
