+++ test/core/image_processing/convolve_2d.cpp
@@ -1,12 +1,13 @@
 //
 // Copyright 2019 Miral Shah <miralshah2211@gmail.com>
+// Copyright 2019-2021 Pranam Lashkari <plashkari628@gmail.com>
 //
 // Use, modification and distribution are subject to the Boost Software License,
 // Version 1.0. (See accompanying file LICENSE_1_0.txt or copy at
 // http://www.boost.org/LICENSE_1_0.txt)
 //
 #include <boost/gil.hpp>
-#include <boost/gil/extension/numeric/convolve.hpp>
+#include <boost/gil/image_processing/convolve.hpp>
 
 #include <boost/core/lightweight_test.hpp>
